move build plan specifc functions to build plan service
