+++ src/main/java/de/tum/in/www1/artemis/service/connectors/jenkins/JenkinsBuildPlanService.java
@@ -1,8 +1,13 @@
 package de.tum.in.www1.artemis.service.connectors.jenkins;
 
+import static de.tum.in.www1.artemis.config.Constants.ASSIGNMENT_REPO_NAME;
+
 import java.io.IOException;
+import java.net.URI;
 import java.net.URL;
 import java.util.stream.Collectors;
+
+import javax.xml.transform.TransformerException;
 
 import org.apache.http.HttpStatus;
 import org.apache.http.client.HttpResponseException;
@@ -11,22 +16,33 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 import org.w3c.dom.Document;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.offbytwo.jenkins.JenkinsServer;
 
 import de.tum.in.www1.artemis.domain.*;
 import de.tum.in.www1.artemis.domain.enumeration.ProgrammingLanguage;
+import de.tum.in.www1.artemis.domain.participation.ProgrammingExerciseParticipation;
 import de.tum.in.www1.artemis.exception.JenkinsException;
+import de.tum.in.www1.artemis.repository.ProgrammingExerciseRepository;
 import de.tum.in.www1.artemis.repository.UserRepository;
 import de.tum.in.www1.artemis.service.connectors.ContinuousIntegrationService;
+import de.tum.in.www1.artemis.service.connectors.jenkins.dto.TestResultsDTO;
 import de.tum.in.www1.artemis.service.connectors.jenkins.jobs.JenkinsJobPermissionsService;
 import de.tum.in.www1.artemis.service.connectors.jenkins.jobs.JenkinsJobService;
+import de.tum.in.www1.artemis.service.util.XmlFileUtils;
 
 @Service
 @Profile("jenkins")
@@ -52,14 +68,18 @@ public class JenkinsBuildPlanService {
 
     private final UserRepository userRepository;
 
+    private final ProgrammingExerciseRepository programmingExerciseRepository;
+
     public JenkinsBuildPlanService(@Qualifier("jenkinsRestTemplate") RestTemplate restTemplate, JenkinsServer jenkinsServer, JenkinsBuildPlanCreator jenkinsBuildPlanCreator,
-            JenkinsJobService jenkinsJobService, JenkinsJobPermissionsService jenkinsJobPermissionsService, UserRepository userRepository) {
+            JenkinsJobService jenkinsJobService, JenkinsJobPermissionsService jenkinsJobPermissionsService, UserRepository userRepository,
+            ProgrammingExerciseRepository programmingExerciseRepository) {
         this.restTemplate = restTemplate;
         this.jenkinsServer = jenkinsServer;
         this.jenkinsBuildPlanCreator = jenkinsBuildPlanCreator;
         this.jenkinsJobService = jenkinsJobService;
         this.userRepository = userRepository;
         this.jenkinsJobPermissionsService = jenkinsJobPermissionsService;
+        this.programmingExerciseRepository = programmingExerciseRepository;
     }
 
     /**
@@ -101,18 +121,95 @@ public class JenkinsBuildPlanService {
     }
 
     /**
+     * Creates a new build plan, configures it for the specified participations and saves it in Jenkins.
+     * The plan is enabled after it has been saved.
+     *
+     * @param participation the programming exercise participation
+     */
+    public void configureBuildPlanForParticipation(ProgrammingExerciseParticipation participation) {
+        // Refetch the programming exercise with the template participation and assign it to programmingExerciseParticipation to make sure it is initialized (and not a proxy)
+        final var programmingExercise = programmingExerciseRepository.findWithTemplateAndSolutionParticipationById(participation.getProgrammingExercise().getId()).get();
+        participation.setProgrammingExercise(programmingExercise);
+
+        final var projectKey = programmingExercise.getProjectKey();
+        final var planKey = participation.getBuildPlanId();
+        final var templateRepoUrl = programmingExercise.getTemplateRepositoryUrl();
+        updateBuildPlanRepositories(projectKey, planKey, ASSIGNMENT_REPO_NAME, participation.getRepositoryUrl(), templateRepoUrl);
+        enablePlan(projectKey, planKey);
+    }
+
+    /**
+     * Updates the repositories that are configured within the build plan with the specified new values.
+     *
+     * @param buildProjectKey the project key of the programming exercise
+     * @param buildPlanKey the build plan id of the participation
+     * @param ciRepoName the name of the repo
+     * @param newRepoUrl the repository url that will replace the old url
+     * @param existingRepoUrl the old repository url that will be replaced
+     */
+    public void updateBuildPlanRepositories(String buildProjectKey, String buildPlanKey, String ciRepoName, String newRepoUrl, String existingRepoUrl) {
+        // remove potential username from repo URL. Jenkins uses the Artemis Admin user and will fail if other usernames are in the URL
+        final var repoUrl = newRepoUrl.replaceAll("(https?://)(.*@)(.*)", "$1$3");
+        final Document jobConfig = jenkinsJobService.getJobConfigForJobInFolder(buildProjectKey, buildPlanKey);
+
+        try {
+            JenkinsBuildPlanUtils.replaceScriptParameters(jobConfig, repoUrl, existingRepoUrl);
+        }
+        catch (IllegalArgumentException e) {
+            log.debug("Pipeline Script not found");
+            log.info("Falling back to old Jenkins setup replacement for build xml");
+            JenkinsBuildPlanUtils.replaceRemoteURLs(jobConfig, repoUrl, ciRepoName);
+        }
+
+        final var errorMessage = "Error trying to configure build plan in Jenkins " + buildPlanKey;
+        try {
+            URI uri = JenkinsEndpoints.PLAN_CONFIG.buildEndpoint(serverUrl.toString(), buildProjectKey, buildPlanKey).build(true).toUri();
+
+            final var headers = new HttpHeaders();
+            headers.setContentType(MediaType.APPLICATION_XML);
+
+            String jobXmlString = XmlFileUtils.writeToString(jobConfig);
+            final var entity = new HttpEntity<>(jobXmlString, headers);
+
+            restTemplate.exchange(uri, HttpMethod.POST, entity, String.class);
+        }
+        catch (RestClientException | TransformerException e) {
+            log.error(errorMessage, e);
+            throw new JenkinsException(errorMessage, e);
+        }
+    }
+
+    /**
+     * Returns the build plan key from the specified test results.
+     *
+     * @param testResultsDTO the test results from Jenkins
+     * @return the build plan key
+     */
+    public String getBuildPlanKeyFromTestResults(TestResultsDTO testResultsDTO) throws JsonProcessingException {
+        final var nameParams = testResultsDTO.getFullName().split(" ");
+        /*
+         * Jenkins gives the full name of a job as <FOLDER NAME> » <JOB NAME> <Build Number> E.g. the third build of an exercise (projectKey = TESTEXC) for its solution build
+         * (TESTEXC-SOLUTION) would be: TESTEXC » TESTEXC-SOLUTION #3 ==> This would mean that at index 2, we have the actual job/plan key, i.e. TESTEXC-SOLUTION
+         */
+        if (nameParams.length != 4) {
+            var requestBodyString = new ObjectMapper().writeValueAsString(testResultsDTO);
+            log.error("Can't extract planKey from requestBody! Not a test notification result!: {}", requestBodyString);
+            throw new JenkinsException("Can't extract planKey from requestBody! Not a test notification result!: " + requestBodyString);
+        }
+
+        return nameParams[2];
+    }
+
+    /**
      * Copies a build plan to another.
      *
      * @param sourceProjectKey the source project key
      * @param sourcePlanName the source plan name
      * @param targetProjectKey the target project key
-     * @param targetProjectName the target project name
      * @param targetPlanName the target plan name
-     * @param targetProjectExists if the project exists
      * @return the key of the created build plan
      */
-    public String copyBuildPlan(String sourceProjectKey, String sourcePlanName, String targetProjectKey, String targetProjectName, String targetPlanName,
-            boolean targetProjectExists) {
+    public String copyBuildPlan(String sourceProjectKey, String sourcePlanName, String targetProjectKey, String targetPlanName) {
         final var cleanTargetName = getCleanPlanName(targetPlanName);
         final var sourcePlanKey = sourceProjectKey + "-" + sourcePlanName;
         final var targetPlanKey = targetProjectKey + "-" + cleanTargetName;
