+++ src/main/java/de/tum/in/www1/artemis/service/connectors/jenkins/JenkinsService.java
@@ -1,33 +1,23 @@
 package de.tum.in.www1.artemis.service.connectors.jenkins;
 
-import static de.tum.in.www1.artemis.config.Constants.ASSIGNMENT_REPO_NAME;
-import static de.tum.in.www1.artemis.config.Constants.TEST_REPO_NAME;
-
 import java.io.IOException;
-import java.net.URI;
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import javax.xml.transform.TransformerException;
-
 import org.apache.commons.collections.CollectionUtils;
-import org.apache.http.client.HttpResponseException;
 import org.jsoup.Jsoup;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
-import org.springframework.http.*;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
-import org.w3c.dom.Document;
-import org.w3c.dom.NodeList;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.offbytwo.jenkins.JenkinsServer;
 
 import de.tum.in.www1.artemis.domain.*;
@@ -36,7 +26,6 @@ import de.tum.in.www1.artemis.domain.participation.ProgrammingExerciseParticipat
 import de.tum.in.www1.artemis.exception.ContinuousIntegrationException;
 import de.tum.in.www1.artemis.exception.JenkinsException;
 import de.tum.in.www1.artemis.repository.FeedbackRepository;
-import de.tum.in.www1.artemis.repository.ProgrammingExerciseRepository;
 import de.tum.in.www1.artemis.repository.ProgrammingSubmissionRepository;
 import de.tum.in.www1.artemis.service.BuildLogEntryService;
 import de.tum.in.www1.artemis.service.connectors.AbstractContinuousIntegrationService;
@@ -46,8 +35,6 @@ import de.tum.in.www1.artemis.service.connectors.jenkins.dto.TestCaseDTO;
 import de.tum.in.www1.artemis.service.connectors.jenkins.dto.TestResultsDTO;
 import de.tum.in.www1.artemis.service.connectors.jenkins.jobs.JenkinsJobService;
 import de.tum.in.www1.artemis.service.dto.AbstractBuildResultNotificationDTO;
-import de.tum.in.www1.artemis.service.util.UrlUtils;
-import de.tum.in.www1.artemis.service.util.XmlFileUtils;
 
 @Profile("jenkins")
 @Service
@@ -55,12 +42,8 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
 
     private static final Logger log = LoggerFactory.getLogger(JenkinsService.class);
 
-    private static final String PIPELINE_SCRIPT_DETECTION_COMMENT = "// ARTEMIS: JenkinsPipeline";
-
     @Value("${jenkins.use-crumb:#{true}}")
     private boolean useCrumb;
-
-    private final ProgrammingExerciseRepository programmingExerciseRepository;
 
     private final JenkinsBuildPlanService jenkinsBuildPlanService;
 
@@ -71,11 +54,9 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
     private final JenkinsInternalUrlService jenkinsInternalUrlService;
 
     public JenkinsService(@Qualifier("jenkinsRestTemplate") RestTemplate restTemplate, JenkinsServer jenkinsServer, ProgrammingSubmissionRepository programmingSubmissionRepository,
-            ProgrammingExerciseRepository programmingExerciseRepository, FeedbackRepository feedbackRepository,
-            @Qualifier("shortTimeoutJenkinsRestTemplate") RestTemplate shortTimeoutRestTemplate, BuildLogEntryService buildLogService,
+            FeedbackRepository feedbackRepository, @Qualifier("shortTimeoutJenkinsRestTemplate") RestTemplate shortTimeoutRestTemplate, BuildLogEntryService buildLogService,
             JenkinsBuildPlanService jenkinsBuildPlanService, JenkinsJobService jenkinsJobService, JenkinsInternalUrlService jenkinsInternalUrlService) {
         super(programmingSubmissionRepository, feedbackRepository, buildLogService, restTemplate, shortTimeoutRestTemplate);
-        this.programmingExerciseRepository = programmingExerciseRepository;
         this.jenkinsServer = jenkinsServer;
         this.jenkinsBuildPlanService = jenkinsBuildPlanService;
         this.jenkinsJobService = jenkinsJobService;
@@ -108,7 +89,7 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
     @Override
     public String copyBuildPlan(String sourceProjectKey, String sourcePlanName, String targetProjectKey, String targetProjectName, String targetPlanName,
             boolean targetProjectExists) {
-        return jenkinsBuildPlanService.copyBuildPlan(sourceProjectKey, sourcePlanName, targetProjectKey, targetProjectName, targetPlanName, targetProjectExists);
+        return jenkinsBuildPlanService.copyBuildPlan(sourceProjectKey, sourcePlanName, targetProjectKey, targetPlanName);
     }
 
     @Override
@@ -118,14 +99,7 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
 
     @Override
     public void configureBuildPlan(ProgrammingExerciseParticipation participation) {
-        // Refetch the programming exercise with the template participation and assign it to programmingExerciseParticipation to make sure it is initialized (and not a proxy)
-        final var programmingExercise = programmingExerciseRepository.findWithTemplateAndSolutionParticipationById(participation.getProgrammingExercise().getId()).get();
-        participation.setProgrammingExercise(programmingExercise);
-        final var projectKey = programmingExercise.getProjectKey();
-        final var planKey = participation.getBuildPlanId();
-        final var templateRepoUrl = programmingExercise.getTemplateRepositoryUrl();
-        updatePlanRepository(projectKey, planKey, ASSIGNMENT_REPO_NAME, null /* not needed */, participation.getRepositoryUrl(), templateRepoUrl, Optional.empty());
-        enablePlan(projectKey, planKey);
+        jenkinsBuildPlanService.configureBuildPlanForParticipation(participation);
     }
 
     @Override
@@ -133,126 +107,7 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
             Optional<List<String>> optionalTriggeredByRepositories) {
         newRepoUrl = jenkinsInternalUrlService.toInternalVcsUrl(newRepoUrl);
         existingRepoUrl = jenkinsInternalUrlService.toInternalVcsUrl(existingRepoUrl);
-
-        // remove potential username from repo URL. Jenkins uses the Artemis Admin user and will fail if other usernames are in the URL
-        final var repoUrl = newRepoUrl.replaceAll("(https?://)(.*@)(.*)", "$1$3");
-        final var jobXmlDocument = jenkinsJobService.getJobConfigForJobInFolder(buildProjectKey, buildPlanKey);
-
-        try {
-            replaceScriptParameters(jobXmlDocument, ciRepoName, repoUrl, existingRepoUrl);
-        }
-        catch (IllegalArgumentException e) {
-            log.info("Falling back to old Jenkins setup replacement for build xml");
-            replaceRemoteURLs(jobXmlDocument, repoUrl, ciRepoName);
-        }
-
-        final var errorMessage = "Error trying to configure build plan in Jenkins " + buildPlanKey;
-        try {
-            URI uri = Endpoint.PLAN_CONFIG.buildEndpoint(serverUrl.toString(), buildProjectKey, buildPlanKey).build(true).toUri();
-
-            final var headers = new HttpHeaders();
-            headers.setContentType(MediaType.APPLICATION_XML);
-
-            String jobXmlString = XmlFileUtils.writeToString(jobXmlDocument);
-            final var entity = new HttpEntity<>(jobXmlString, headers);
-
-            restTemplate.exchange(uri, HttpMethod.POST, entity, String.class);
-        }
-        catch (RestClientException | TransformerException e) {
-            log.error(errorMessage, e);
-            throw new JenkinsException(errorMessage, e);
-        }
-    }
-
-    private void replaceScriptParameters(Document jobXmlDocument, String ciRepoName, String repoUrl, String baseRepoUrl) throws IllegalArgumentException {
-        final var scriptNode = findScriptNode(jobXmlDocument);
-        if (scriptNode == null || scriptNode.getFirstChild() == null) {
-            log.debug("Pipeline Script not found");
-            throw new IllegalArgumentException("Pipeline Script not found");
-        }
-
-        String pipeLineScript = scriptNode.getFirstChild().getTextContent().trim();
-        // If the script does not start with "pipeline" or the special comment,
-        // it is not actually a pipeline script, but a deprecated programming exercise with an old build xml configuration
-        if (!pipeLineScript.startsWith("pipeline") && !pipeLineScript.startsWith(PIPELINE_SCRIPT_DETECTION_COMMENT)) {
-            log.debug("Pipeline Script not found");
-            throw new IllegalArgumentException("Pipeline Script not found");
-        }
-        // Replace repo URL
-        // TODO: properly replace the baseRepoUrl with repoUrl by looking up the ciRepoName in the pipelineScript
-        pipeLineScript = pipeLineScript.replace(baseRepoUrl, repoUrl);
-
-        scriptNode.getFirstChild().setTextContent(pipeLineScript);
-    }
-
-    /**
-     * Replace old XML files that are not based on pipelines.
-     * Will be removed in the future
-     *
-     * @param jobXmlDocument the Document where the remote config should replaced
-     */
-    @Deprecated
-    private void replaceRemoteURLs(Document jobXmlDocument, String repoUrl, String repoNameInCI) throws IllegalArgumentException {
-        final var remoteUrlNode = findUserRemoteConfigFor(jobXmlDocument, repoNameInCI);
-        if (remoteUrlNode == null || remoteUrlNode.getFirstChild() == null) {
-            throw new IllegalArgumentException("Url to replace not found in job xml document");
-        }
-        remoteUrlNode.getFirstChild().setNodeValue(repoUrl);
-    }
-
-    private org.w3c.dom.Node findScriptNode(Document jobXmlDocument) {
-        final var userRemoteConfigs = jobXmlDocument.getElementsByTagName("script");
-        return userRemoteConfigs.item(0);
-    }
-
-    private org.w3c.dom.Node findUserRemoteConfigFor(Document jobXmlDocument, String repoNameInCI) {
-        final var userRemoteConfigs = jobXmlDocument.getElementsByTagName("hudson.plugins.git.UserRemoteConfig");
-        if (userRemoteConfigs.getLength() != 2) {
-            throw new IllegalArgumentException("Configuration of build plans currently only supports a model with two repositories, ASSIGNMENT and TESTS");
-        }
-        var firstUserRemoteConfig = userRemoteConfigs.item(0).getChildNodes();
-        var urlElement = findUrlElement(firstUserRemoteConfig, repoNameInCI);
-        if (urlElement != null) {
-            return urlElement;
-        }
-        var secondUserRemoteConfig = userRemoteConfigs.item(1).getChildNodes();
-        urlElement = findUrlElement(secondUserRemoteConfig, repoNameInCI);
-        return urlElement;
-    }
-
-    private org.w3c.dom.Node findUrlElement(NodeList nodeList, String repoNameInCI) {
-        boolean found = false;
-        org.w3c.dom.Node urlNode = null;
-        for (int i = 0; i < nodeList.getLength(); i++) {
-            var childElement = nodeList.item(i);
-            if ("name".equalsIgnoreCase(childElement.getNodeName())) {
-                var nameValue = childElement.hasChildNodes() ? childElement.getFirstChild().getNodeValue() : null;
-                // this name was added recently, so we cannot assume that all job xml files include this name
-                if (repoNameInCI.equalsIgnoreCase(nameValue)) {
-                    found = true;
-                }
-            }
-            else if ("url".equalsIgnoreCase(childElement.getNodeName())) {
-                urlNode = childElement;
-                if (!found) {
-                    // fallback for old xmls
-                    var urlValue = childElement.hasChildNodes() ? childElement.getFirstChild().getNodeValue() : null;
-                    if (urlValue != null && repoNameInCI.equals(ASSIGNMENT_REPO_NAME) && ((urlValue.contains("-exercise.git") || (urlValue.contains("-solution.git"))))) {
-                        found = true;
-                    }
-                    else if (urlValue != null && repoNameInCI.equals(TEST_REPO_NAME) && urlValue.contains("-tests.git")) {
-                        found = true;
-                    }
-                }
-            }
-        }
-
-        if (found && urlNode != null) {
-            return urlNode;
-        }
-        else {
-            return null;
-        }
+        jenkinsBuildPlanService.updateBuildPlanRepositories(buildProjectKey, buildPlanKey, ciRepoName, newRepoUrl, existingRepoUrl);
     }
 
     @Override
@@ -264,20 +119,7 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
 
     @Override
     public void deleteProject(String projectKey) {
-        try {
-            jenkinsServer.deleteJob(projectKey, useCrumb);
-        }
-        catch (HttpResponseException e) {
-            // We don't throw an exception if the project doesn't exist in Jenkins (404 status)
-            if (e.getStatusCode() != org.apache.http.HttpStatus.SC_NOT_FOUND) {
-                log.error(e.getMessage(), e);
-                throw new JenkinsException("Error while trying to delete folder in Jenkins for " + projectKey, e);
-            }
-        }
-        catch (IOException e) {
-            log.error(e.getMessage(), e);
-            throw new JenkinsException("Error while trying to delete folder in Jenkins for " + projectKey, e);
-        }
+        jenkinsJobService.deleteJob(projectKey);
     }
 
     @Override
@@ -287,19 +129,8 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
 
     @Override
     public String getPlanKey(Object requestBody) throws Exception {
-        final var result = TestResultsDTO.convert(requestBody);
-        final var nameParams = result.getFullName().split(" ");
-        /*
-         * Jenkins gives the full name of a job as <FOLDER NAME> » <JOB NAME> <Build Number> E.g. the third build of an exercise (projectKey = TESTEXC) for its solution build
-         * (TESTEXC-SOLUTION) would be: TESTEXC » TESTEXC-SOLUTION #3 ==> This would mean that at index 2, we have the actual job/plan key, i.e. TESTEXC-SOLUTION
-         */
-        if (nameParams.length != 4) {
-            var requestBodyString = new ObjectMapper().writeValueAsString(requestBody);
-            log.error("Can't extract planKey from requestBody! Not a test notification result!: {}", requestBodyString);
-            throw new JenkinsException("Can't extract planKey from requestBody! Not a test notification result!: " + requestBodyString);
-        }
-
-        return nameParams[2];
+        final TestResultsDTO testResultsDTO = TestResultsDTO.convert(requestBody);
+        return jenkinsBuildPlanService.getBuildPlanKeyFromTestResults(testResultsDTO);
     }
 
     @Override
@@ -510,12 +341,12 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
 
     @Override
     public void giveProjectPermissions(String projectKey, List<String> groups, List<CIPermission> permissions) {
-        // TODO after decision on how to handle users on Jenkins has been made
+        // Not needed since Jenkins doesn't support project permissions
     }
 
     @Override
     public void removeAllDefaultProjectPermissions(String projectKey) {
-        // TODO after decision on how to handle users on Jenkins has been made
+        // Not needed since Jenkins doesn't support project permissions
     }
 
     @Override
@@ -540,25 +371,6 @@ public class JenkinsService extends AbstractContinuousIntegrationService {
         catch (IOException e) {
             log.error(e.getMessage(), e);
             throw new JenkinsException("Error creating folder for exercise " + programmingExercise, e);
-        }
-    }
-
-    private enum Endpoint {
-
-        NEW_PLAN("job", "<projectKey>", "createItem"), NEW_FOLDER("createItem"), DELETE_FOLDER("job", "<projectKey>", "doDelete"),
-        DELETE_JOB("job", "<projectKey>", "job", "<planName>", "doDelete"), PLAN_CONFIG("job", "<projectKey>", "job", "<planKey>", "config.xml"),
-        TRIGGER_BUILD("job", "<projectKey>", "job", "<planKey>", "build"), ENABLE("job", "<projectKey>", "job", "<planKey>", "enable"),
-        TEST_RESULTS("job", "<projectKey>", "job", "<planKey>", "lastBuild", "testResults", "api", "json"),
-        LAST_BUILD("job", "<projectKey>", "job", "<planKey>", "lastBuild", "api", "json");
-
-        private final List<String> pathSegments;
-
-        Endpoint(String... pathSegments) {
-            this.pathSegments = Arrays.asList(pathSegments);
-        }
-
-        public UriComponentsBuilder buildEndpoint(String baseUrl, Object... args) {
-            return UrlUtils.buildEndpoint(baseUrl, pathSegments, args);
         }
     }
 }
